+++ src/todo/decorators/todo_done.ts
@@ -1,13 +1,14 @@
 /* IMPORT */
 
 import * as vscode from 'vscode';
+import Config from '../../config';
 import Consts from '../../consts';
 import TodoDoneItem from '../items/todo_done';
 import Line from './line';
 
 /* DECORATION TYPES */
 
-const TODO_DONE = vscode.window.createTextEditorDecorationType({
+const TODO_DONE_OPTIONS = {
     color: Consts.colors.done,
     rangeBehavior: vscode.DecorationRangeBehavior.ClosedOpen,
     dark: {
@@ -16,12 +17,19 @@ const TODO_DONE = vscode.window.createTextEditorDecorationType({
     light: {
         color: Consts.colors.light.done,
     },
+};
+
+const TODO_DONE = vscode.window.createTextEditorDecorationType(TODO_DONE_OPTIONS);
+
+const TODO_DONE_STRIKETHROUGH = vscode.window.createTextEditorDecorationType({
+    ...TODO_DONE_OPTIONS,
+    textDecoration: 'line-through',
 });
 
 /* TODO DONE */
 
 class TodoDone extends Line {
-    TYPES = [TODO_DONE];
+    TYPES = [TODO_DONE, TODO_DONE_STRIKETHROUGH];
 
     getItemRanges(todoDone: TodoDoneItem, negRange?: vscode.Range | vscode.Range[]) {
         return [
@@ -30,6 +38,22 @@ class TodoDone extends Line {
                 todoDone.range,
                 negRange || [Consts.regexes.tag, Consts.regexes.formattedCode]
             ),
+        ];
+    }
+
+    getDecorations(todosDone: TodoDoneItem[], negRange?: vscode.Range | vscode.Range[]) {
+        const ranges = this.getItemsRanges(todosDone, negRange)[0] || [],
+            strikethrough = Config.getKey('decorations.done.strikethrough') !== false;
+
+        return [
+            {
+                type: TODO_DONE,
+                ranges: strikethrough ? [] : ranges,
+            },
+            {
+                type: TODO_DONE_STRIKETHROUGH,
+                ranges: strikethrough ? ranges : [],
+            },
         ];
     }
 }